// src/app/types/weatherGov.d.ts
export interface paths {
// ...
    "/alerts/active": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Returns all currently active alerts */
        get: operations["alerts_active"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Alert: {
            id?: components["schemas"]["AlertId"];
            /** @description A textual description of the area affected by the alert. */
            areaDesc?: string;
            /** @description Lists of codes for NWS public zones and counties affected by the alert. */
            geocode?: {
                /** @description A list of NWS public zone or county identifiers. */
                UGC?: components["schemas"]["NWSZoneID"][];
                /** @description A list of SAME (Specific Area Message Encoding) codes for affected counties. */
                SAME?: string[];
            };
            /** @description An array of API links for zones affected by the alert. This is an API-specific extension field and is not part of the CAP specification.
             *      */
            affectedZones?: string[];
            /** @description A list of prior alerts that this alert updates or replaces. */
            references?: {
                /**
                 * Format: uri
                 * @description An API link to the prior alert.
                 */
                "@id"?: string;
                identifier?: components["schemas"]["AlertId"];
                /** @description The sender of the prior alert. */
                sender?: string;
                /**
                 * Format: date-time
                 * @description The time the prior alert was sent.
                 */
                sent?: string;
            }[];
            /**
             * Format: date-time
             * @description The time of the origination of the alert message.
             */
            sent?: string;
            /**
             * Format: date-time
             * @description The effective time of the information of the alert message.
             */
            effective?: string;
            /**
             * Format: date-time
             * @description The expected time of the beginning of the subject event of the alert message.
             */
            onset?: string | null;
            /**
             * Format: date-time
             * @description The expiry time of the information of the alert message.
             */
            expires?: string;
            /**
             * Format: date-time
             * @description The expected end time of the subject event of the alert message.
             */
            ends?: string | null;
            status?: components["schemas"]["AlertStatus"];
            messageType?: components["schemas"]["AlertMessageType"];
            /**
             * @description The code denoting the category of the subject event of the alert message.
             * @enum {string}
             */
            category?: "Met" | "Geo" | "Safety" | "Security" | "Rescue" | "Fire" | "Health" | "Env" | "Transport" | "Infra" | "CBRNE" | "Other";
            severity?: components["schemas"]["AlertSeverity"];
            certainty?: components["schemas"]["AlertCertainty"];
            urgency?: components["schemas"]["AlertUrgency"];
            /** @description The text denoting the type of the subject event of the alert message. */
            event?: string;
            /** @description Email address of the NWS webmaster. */
            sender?: string;
            /** @description The text naming the originator of the alert message. */
            senderName?: string;
            /** @description The text headline of the alert message. */
            headline?: string | null;
            /** @description The text describing the subject event of the alert message. */
            description?: string;
            /** @description The text describing the recommended action to be taken by recipients of the alert message.
             *      */
            instruction?: string | null;
            /**
             * @description The code denoting the type of action recommended for the target audience.
             *     This corresponds to responseType in the CAP specification.
             *
             * @enum {string}
             */
            response?: "Shelter" | "Evacuate" | "Prepare" | "Execute" | "Avoid" | "Monitor" | "Assess" | "AllClear" | "None";
            /** @description System-specific additional parameters associated with the alert message.
             *     The keys in this object correspond to parameter definitions in the NWS CAP specification.
             *      */
            parameters?: {
                [key: string]: unknown[];
            };
        };
        AlertCollection: {
            /** @description A title describing the alert collection */
            title?: string;
            /**
             * Format: date-time
             * @description The last time a change occurred to this collection
             */
            updated?: string;
            pagination?: components["schemas"]["PaginationInfo"];
        };
        AlertCollectionGeoJson: components["schemas"]["GeoJsonFeatureCollection"] & {
            features?: {
                properties?: components["schemas"]["Alert"];
            }[];
        } & components["schemas"]["AlertCollection"];
// ...
    };
    responses: {
        /** @description A collection of alerts. */
        AlertCollection: {
            headers: {
                "X-Correlation-Id": components["headers"]["CorrelationId"];
                "X-Request-Id": components["headers"]["RequestId"];
                "X-Server-Id": components["headers"]["ServerId"];
                [name: string]: unknown;
            };
            content: {
                "application/geo+json": components["schemas"]["AlertCollectionGeoJson"];
                "application/ld+json": components["schemas"]["AlertCollectionJsonLd"];
                "application/atom+xml": components["schemas"]["AlertAtomFeed"];
            };
        };
        /** @description An error response. */
        Error: {
            headers: {
                "X-Correlation-Id": components["headers"]["CorrelationId"];
                "X-Request-Id": components["headers"]["RequestId"];
                "X-Server-Id": components["headers"]["ServerId"];
                [name: string]: unknown;
            };
            content: {
                "application/problem+json": components["schemas"]["ProblemDetail"];
            };
        };
        /** @description A forecast for a gridpoint. */
        GridpointForecast: {
            headers: {
                "X-Correlation-Id": components["headers"]["CorrelationId"];
                "X-Request-Id": components["headers"]["RequestId"];
                "X-Server-Id": components["headers"]["ServerId"];
                [name: string]: unknown;
            };
            content: {
                "application/geo+json": components["schemas"]["GridpointForecastGeoJson"];
                "application/ld+json": components["schemas"]["GridpointForecastJsonLd"];
                "application/vnd.noaa.dwml+xml": unknown;
            };
        };
        /** @description An observation record. */
        Observation: {
            headers: {
                "X-Correlation-Id": components["headers"]["CorrelationId"];
                "X-Request-Id": components["headers"]["RequestId"];
                "X-Server-Id": components["headers"]["ServerId"];
                [name: string]: unknown;
            };
            content: {
                "application/geo+json": components["schemas"]["ObservationGeoJson"];
                "application/ld+json": components["schemas"]["ObservationJsonLd"];
                "application/vnd.noaa.uswx+xml": unknown;
                "application/vnd.noaa.obs+xml": unknown;
            };
        };
        /** @description A collection of observation records. */
        ObservationCollection: {
            headers: {
                "X-Correlation-Id": components["headers"]["CorrelationId"];
                "X-Request-Id": components["headers"]["RequestId"];
                "X-Server-Id": components["headers"]["ServerId"];
                [name: string]: unknown;
            };
            content: {
                "application/geo+json": components["schemas"]["ObservationCollectionGeoJson"];
                "application/ld+json": components["schemas"]["ObservationCollectionJsonLd"];
            };
        };
        /** @description A collection of observation stations. */
        ObservationStationCollection: {
            headers: {
                "X-Correlation-Id": components["headers"]["CorrelationId"];
                "X-Request-Id": components["headers"]["RequestId"];
                "X-Server-Id": components["headers"]["ServerId"];
                [name: string]: unknown;
            };
            content: {
                "application/geo+json": components["schemas"]["ObservationStationCollectionGeoJson"];
                "application/ld+json": components["schemas"]["ObservationStationCollectionJsonLd"];
            };
        };
    };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
    alerts_query: {
// ...
        responses: {
            200: components["responses"]["AlertCollection"];
            /** @description Certain common queries may be redirected to discrete URLs */
            301: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            default: components["responses"]["Error"];
        };
    };
// ...
}